Reorganize to have a standalone siav2_query() interface
